backend/pyproject.toml
[project]
name = "expenses-backend"
version = "0.1.0"
requires-python = ">=3.10,<3.13"
dependencies = [
  "fastapi>=0.100.0",
  "uvicorn[standard]>=0.20.0",
  "sqlalchemy>=2.0.0",
  "aiosqlite>=0.19.0",
  "alembic>=1.12.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.0.0",
  "python-dotenv>=1.0.0",
  "python-jose[cryptography]>=3.3.0",
  "argon2-cffi>=23.1.0",
]
